requirements.txt
chart_studio==1.1.0
pandas==1.5.3
Pillow==9.4.0
plotly==4.14.3
streamlit==1.18.1
